add daily expenses fragment and more

+++ app/src/main/java/com/unex/expenses/models/Spending.kt
@@ -6,7 +6,7 @@ import java.util.*
 
 @Entity(tableName = "spendings")
 class Spending(
-        private val amount: Double,
+        private val amount: Int,
         private val date: Date,
         private val tags: Set<String>,
         private val description: String?
